feat: refactor bank section
